refactor: api module added, 'dto' package is common
