+++ api/src/main/java/weather/api/dto/TemperatureDto.java
@@ -1,6 +1,9 @@
-package weather.dto;
+package weather.api.dto;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NonNull;
 
 @Data
 @Builder
