+++ pom.xml
@@ -3,6 +3,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
+    <parent>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-starter-parent</artifactId>
+        <version>3.2.4</version>
+        <relativePath/>
+    </parent>
+
     <groupId>weather</groupId>
     <artifactId>weather</artifactId>
     <version>0.0.1-SNAPSHOT</version>
@@ -14,6 +21,7 @@
     <modules>
         <module>service/reporter</module>
         <module>service/saver</module>
+        <module>api</module>
     </modules>
 
     <properties>
