+++ service/reporter/src/main/java/weather/mapper/WeatherResponseMapper.java
@@ -3,7 +3,7 @@ package weather.mapper;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import weather.dto.WeatherDto;
+import weather.api.dto.WeatherDto;
 import weather.model.Weather;
 import weather.response.TemperatureResponse;
 import weather.response.WeatherResponse;
