+++ service/reporter/src/test/resources/application.yaml
@@ -7,8 +7,3 @@ security:
 integration:
   openweather:
     api-key: apiKey
-
-spring:
-  kafka:
-    consumer:
-      group-id: test
