+++ service/saver/src/main/java/weather/WeatherSaverApplication.java
@@ -1,0 +1,12 @@
+package weather;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WeatherSaverApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WeatherSaverApplication.class, args);
+    }
+}
