+++ service/saver/src/main/java/weather/configuration/kafka/KafkaConsumerConfiguration.java
