+++ service/saver/src/main/java/weather/configuration/kafka/KafkaProperties.java
@@ -1,5 +1,6 @@
 package weather.configuration.kafka;
 
+import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
@@ -14,8 +15,19 @@ import java.util.List;
 @Validated
 @Getter
 @Setter
-class KafkaProperties {
+public class KafkaProperties {
 
     @NotNull(message = "spring.kafka.bootstrap-servers property is not specified")
     private List<String> bootstrapServers;
+    @Valid
+    private Consumer consumer = new Consumer();
+
+    @Validated
+    @Getter
+    @Setter
+    static class Consumer {
+
+        @NotNull(message = "spring.kafka.consumer.group-id property is not specified")
+        private String groupId;
+    }
 }
