+++ service/saver/src/main/resources/application.yaml
@@ -1,0 +1,6 @@
+spring:
+  kafka:
+    bootstrap-servers:
+      - localhost:9094
+    consumer:
+      group-id: weather
